apps/comment_labeler/pyproject.toml
[project]
name = "comment-labeler"
version = "0.1.0"
description = "Tools for creating training data for horse racing comment analysis"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "dash>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
